requirements.txt
igdb-api-v4
pymongo
tqdm